Fixed issue #779: Utils/tests/Expect tests adopted for nnbd.

+++ Utils/tests/Expect/approxEquals_A01_t01.dart
@@ -5,20 +5,21 @@
  */
 /**
  * @assertion static void approxEquals(num expected, num actual, [num tolerance = null, String reason = null])
- * Failure if the difference between expected and actual is greater than the given tolerance. 
- * @description Checks that no exception is thrown when the difference between arguments is equal 
- * to specified non-negative, non-NAN tolerance, regardless of reason (the last argument). Also
- * checks the case when the specified tolerance is negative zero (0.0 == -0.0).
+ * Failure if the difference between expected and actual is greater than the
+ * given tolerance.
+ * @description Checks that no exception is thrown when the difference between
+ * arguments is equal to specified non-negative, non-NAN tolerance, regardless
+ * of reason (the last argument). Also checks the case when the specified
+ * tolerance is negative zero (0.0 == -0.0).
  * @author rodionov
- * @reviewer varlax
  */
 import "../../../Utils/expect.dart";
 import "dart:math" as Math;
 
 main() {
-  final double MIN_DOUBLE = Math.pow(2.0, -1074);
+  final double MIN_DOUBLE = Math.pow(2.0, -1074) as double;
   final double NEG_MIN_DOUBLE = -1 * MIN_DOUBLE; 
-  final double MAX_DOUBLE = (2 - Math.pow(2.0, -52)) * Math.pow(2.0, 1023);
+  final double MAX_DOUBLE = (2 - Math.pow(2.0, -52)) * Math.pow(2.0, 1023) as double;
   final double NEG_MAX_DOUBLE = -1 * MAX_DOUBLE; 
 
   Expect.approxEquals(100.0, 101, 1);
