+++ Utils/tests/Expect/approxEquals_A01_t04.dart
@@ -5,21 +5,22 @@
  */
 /**
  * @assertion static void approxEquals(num expected, num actual, [num tolerance = null, String reason = null])
- * Failure if the difference between expected and actual is greater than the given tolerance. 
- * @description Checks that no exception is thrown when the absolute difference between arguments 
- * is strictly less than specified non-negative, non-NAN tolerance, regardless of reason (the last argument).
+ * Failure if the difference between expected and actual is greater than the
+ * given tolerance.
+ * @description Checks that no exception is thrown when the absolute difference
+ * between arguments is strictly less than specified non-negative, non-NAN
+ * tolerance, regardless of reason (the last argument).
  * @author rodionov
  * @author varlax
- * @reviewer varlax
  */
 import "../../../Utils/expect.dart";
 
 import "dart:math" as Math;
 
 main() {
-  final double MIN_DOUBLE = Math.pow(2.0, -1074);
+  final double MIN_DOUBLE = Math.pow(2.0, -1074) as double;
   final double NEG_MIN_DOUBLE = -1 * MIN_DOUBLE; 
-  final double MAX_DOUBLE = (2 - Math.pow(2.0, -52)) * Math.pow(2.0, 1023);
+  final double MAX_DOUBLE = (2 - Math.pow(2.0, -52)) * Math.pow(2.0, 1023) as double;
   final double NEG_MAX_DOUBLE = -1 * MAX_DOUBLE; 
 
   Expect.approxEquals(100.0, 101, 1.1);
